fix(lang_math): the Integer cache index is unbounded above, so widen before subtracting
`canonical_wrapper_if_cached`'s `"I"` arm guards only the LOW end
(`x >= INTEGER_CACHE_LOW`) — deliberately, because the backing store's
length is the high bound and `read`'s `get(idx)` is the test. But the
index it computes is `x - INTEGER_CACHE_LOW` in `i32`, so
`Value::Int(i32::MAX)` overflows and a debug build panics where the
contract says MISS.

That is not hypothetical: `canonical_wrapper_if_cached_follows_the_
configured_integer_bound` asserts exactly `Int(i32::MAX) -> None` and
has been failing `cargo test -p cratonvm-native-builtins --lib` on dev.
Release builds wrap instead and land on an out-of-range `get`, which
misses — so only the debug gate was red, and only the gate.

Both store-SIZING sites in this file already widen to `i64` for the same
reason; this is the read side catching up. Found while running the
crate's suite for an unrelated TLS change, not by looking for it.

+++ native-builtins/src/lang_math.rs
@@ -3477,7 +3477,16 @@ pub fn canonical_wrapper_if_cached(
         // instead would need the memo, and a VM that has not boxed an `int`
         // yet has neither — which is a miss either way.
         ("I", Value::Int(x)) if x >= INTEGER_CACHE_LOW => {
-            read(integer_cache(), vm_identity, (x - INTEGER_CACHE_LOW) as usize)
+            // Widen before subtracting: `x` is unbounded ABOVE here (the
+            // `get(idx)` below is the only bound), so `i32::MAX -
+            // INTEGER_CACHE_LOW` overflows an `i32` and panics a debug build
+            // where it must simply MISS. Both store-SIZING sites in this file
+            // already widen to `i64` for exactly this reason.
+            read(
+                integer_cache(),
+                vm_identity,
+                (x as i64 - INTEGER_CACHE_LOW as i64) as usize,
+            )
         }
         ("J", Value::Long(x)) if (-128..=127).contains(&x) => {
             read(long_cache(), vm_identity, (x + 128) as usize)
